pyproject.toml
[project]
name = "rrms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.13.0",
    "flask>=3.1.3",
    "flask-wtf>=1.2.2",
    "libsql-client>=0.3.1",
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
]
